+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Support validation using `validator` crate
+
 ## [0.3.9] - 2026-02-14
 ### Added
 - Support deprecated config fields reporting via `Config::with_deprecation_clb`